athena/requirements.txt
Flask==3.1.*
firebase-admin==7.4.*
flask-wtf==1.3.*
flask-limiter==4.1.*
gunicorn==25.3.*
google-cloud-firestore==2.27.*
google-cloud-storage==3.10.*
google-cloud-secret-manager==2.27.*
google-cloud-logging>=3.10.0
# OpenTelemetry — distributed tracing → Cloud Trace.
# Versions paired: SDK 1.27 ↔ instrumentation 0.48b0 (OTel's stable+beta
# split).  GCP exporter 1.7.* is the latest line tested against SDK 1.27.
opentelemetry-api==1.27.*
opentelemetry-sdk==1.27.*
opentelemetry-exporter-gcp-trace==1.7.*
opentelemetry-propagator-gcp==1.7.*
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-jinja2==0.48b0
icalendar==7.0.*
vobject==0.9.*
bcrypt==5.0.*
python-dotenv==1.2.*
reportlab==4.4.*
# Pillow is a transitive dependency of reportlab (which only requires >=9.0.0).
# Pinned explicitly here to clear CVE-2023-44271 (Pillow <10), CVE-2026-25990 (PSD OOB write,
# fixed in 12.2.0), and CVE-2026-40192 (FITS GZIP decompression bomb, fixed in 12.2.0).
Pillow==12.*
Markdown==3.10.*
bleach==6.3.*
defusedxml==0.7.*
